requirements.txt
absl-py==2.3.1
attrs==25.4.0
cffi==2.0.0
contourpy==1.3.0
cycler==0.12.1
cysignals==1.12.5
flatbuffers==25.9.23
fonttools==4.60.1
importlib_metadata==8.7.0
importlib_resources==6.5.2
jax==0.4.30
jaxlib==0.4.30
kiwisolver==1.4.7
matplotlib==3.9.4
mediapipe==0.10.21
ml_dtypes==0.5.3
msgpack==1.1.2
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==25.0
pillow==11.3.0
protobuf==4.25.8
pupil-detectors==2.0.2
pupil-labs-uvc==1.0.0
pycparser==2.23
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pyzmq==27.1.0
scipy==1.13.1
sentencepiece==0.2.1
six==1.17.0
sounddevice==0.5.3
websockets==15.0.1
zipp==3.23.0
